+++ mockredis/client.py
@@ -5,9 +5,9 @@ from datetime import datetime, timedelta
 from hashlib import sha1
 from operator import add
 from random import choice, sample
-import time
 import re
 import sys
+import time
 
 from mockredis.clock import SystemClock
 from mockredis.lock import MockRedisLock
@@ -60,10 +60,10 @@ class MockRedis(object):
     #### Connection Functions ####
 
     def echo(self, msg):
-        return msg
+        return self._encode(msg)
 
     def ping(self):
-        return "PONG"
+        return b'PONG'
 
     #### Transactions Functions ####
 
@@ -104,35 +104,36 @@ class MockRedis(object):
     #### Keys Functions ####
 
     def type(self, key):
+        key = self._encode(key)
         if key not in self.redis:
-            return 'none'
+            return b'none'
         type_ = type(self.redis[key])
         if type_ is dict:
-            return 'hash'
+            return b'hash'
         elif type_ is str:
-            return 'string'
+            return b'string'
         elif type_ is set:
-            return 'set'
+            return b'set'
         elif type_ is list:
-            return 'list'
+            return b'list'
         elif type_ is SortedSet:
-            return 'zset'
+            return b'zset'
         raise TypeError("unhandled type {}".format(type_))
 
     def keys(self, pattern='*'):
         """Emulate keys."""
         # Make a regex out of pattern. The only special matching character we look for is '*'
-        regex = '^' + pattern.replace('*', '.*') + '$'
+        regex = re.compile(b'^' + re.escape(self._encode(pattern)).replace(b'\\*', b'.*') + b'$')
 
         # Find every key that matches the pattern
-        result = [key for key in self.redis.keys() if re.match(regex, key)]
+        result = [key for key in self.redis.keys() if regex.match(key)]
 
         return result
 
     def delete(self, *keys):
         """Emulate delete."""
         key_counter = 0
-        for key in map(str, keys):
+        for key in map(self._encode, keys):
             if key in self.redis:
                 del self.redis[key]
                 key_counter += 1
@@ -147,7 +148,7 @@ class MockRedis(object):
 
     def exists(self, key):
         """Emulate exists."""
-        return key in self.redis
+        return self._encode(key) in self.redis
     __contains__ = exists
 
     def _expire(self, key, delta):
@@ -160,15 +161,16 @@ class MockRedis(object):
     def expire(self, key, delta):
         """Emulate expire"""
         delta = delta if isinstance(delta, timedelta) else timedelta(seconds=delta)
-        return self._expire(key, delta)
+        return self._expire(self._encode(key), delta)
 
     def pexpire(self, key, milliseconds):
         """Emulate pexpire"""
-        return self._expire(key, timedelta(milliseconds=milliseconds))
+        return self._expire(self._encode(key), timedelta(milliseconds=milliseconds))
 
     def expireat(self, key, when):
         """Emulate expireat"""
         expire_time = datetime.fromtimestamp(when)
+        key = self._encode(key)
         if key in self.redis:
             self.timeouts[key] = expire_time
             return True
@@ -203,12 +205,12 @@ class MockRedis(object):
         """
         Returns time to live in milliseconds if output_ms is True, else returns seconds.
         """
+        key = self._encode(key)
         if key not in self.redis:
             # as of redis 2.8, -2 is returned if the key does not exist
             return long(-2) if self.strict else None
         if key not in self.timeouts:
             # as of redis 2.8, -1 is returned if the key is persistent
-            # redis-py returns None; command docs say -1
             return long(-1) if self.strict else None
 
         time_to_live = get_total_milliseconds(self.timeouts[key] - self.clock.now())
@@ -237,6 +239,8 @@ class MockRedis(object):
         return 1 if self._rename(old_key, new_key, True) else 0
 
     def _rename(self, old_key, new_key, nx=False):
+        old_key = self._encode(old_key)
+        new_key = self._encode(new_key)
         if old_key in self.redis and (not nx or new_key not in self.redis):
             self.redis[new_key] = self.redis.pop(old_key)
             return True
@@ -245,10 +249,8 @@ class MockRedis(object):
     #### String Functions ####
 
     def get(self, key):
-
-        # Override the default dict
-        result = None if key not in self.redis else self.redis[key]
-        return result
+        key = self._encode(key)
+        return self.redis.get(key)
 
     def __getitem__(self, name):
         """
@@ -273,6 +275,9 @@ class MockRedis(object):
         If px and ex are both set, the preference is given to px.
         If the key is not set for some reason, the lib function returns None.
         """
+        key = self._encode(key)
+        value = self._encode(value)
+
         if nx and xx:
             return None
         mode = "nx" if nx else "xx" if xx else None
@@ -299,7 +304,7 @@ class MockRedis(object):
         return old_value
 
     def _set(self, key, value):
-        self.redis[key] = str(value)
+        self.redis[key] = self._encode(value)
 
         # removing the timeout
         if key in self.timeouts:
@@ -382,8 +387,8 @@ class MockRedis(object):
         else:
             mapping = kwargs
 
-        for key, value in mapping.items():
-            if key in self.redis:
+        for key in mapping.keys():
+            if self._encode(key) in self.redis:
                 return False
         for key, value in mapping.items():
             self.set(key, value)
@@ -391,22 +396,21 @@ class MockRedis(object):
         return True
 
     def decr(self, key, amount=1):
-        """Emulate decr."""
+        key = self._encode(key)
         previous_value = long(self.redis.get(key, '0'))
-        self.redis[key] = str(previous_value - amount)
+        self.redis[key] = self._encode(previous_value - amount)
         return long(self.redis[key])
 
-    def decrby(self, key, amount=1):
-        return self.decr(key, amount)
+    decrby = decr
 
     def incr(self, key, amount=1):
         """Emulate incr."""
+        key = self._encode(key)
         previous_value = long(self.redis.get(key, '0'))
-        self.redis[key] = str(previous_value + amount)
+        self.redis[key] = self._encode(previous_value + amount)
         return long(self.redis[key])
 
-    def incrby(self, key, amount=1):
-        return self.incr(key, amount)
+    incrby = incr
 
     #### Hash Functions ####
 
@@ -414,13 +418,13 @@ class MockRedis(object):
         """Emulate hexists."""
 
         redis_hash = self._get_hash(hashkey, 'HEXISTS')
-        return str(attribute) in redis_hash
+        return self._encode(attribute) in redis_hash
 
     def hget(self, hashkey, attribute):
         """Emulate hget."""
 
         redis_hash = self._get_hash(hashkey, 'HGET')
-        return redis_hash.get(str(attribute))
+        return redis_hash.get(self._encode(attribute))
 
     def hgetall(self, hashkey):
         """Emulate hgetall."""
@@ -434,12 +438,12 @@ class MockRedis(object):
         redis_hash = self._get_hash(hashkey, 'HDEL')
         count = 0
         for key in keys:
-            attribute = str(key)
+            attribute = self._encode(key)
             if attribute in redis_hash:
                 count += 1
                 del redis_hash[attribute]
                 if not redis_hash:
-                    del self.redis[hashkey]
+                    self.delete(hashkey)
         return count
 
     def hlen(self, hashkey):
@@ -452,8 +456,8 @@ class MockRedis(object):
 
         redis_hash = self._get_hash(hashkey, 'HMSET', create=True)
         for key, value in value.items():
-            attribute = str(key)
-            redis_hash[attribute] = str(value)
+            attribute = self._encode(key)
+            redis_hash[attribute] = self._encode(value)
         return True
 
     def hmget(self, hashkey, keys, *args):
@@ -461,26 +465,26 @@ class MockRedis(object):
 
         redis_hash = self._get_hash(hashkey, 'HMGET')
         attributes = self._list_or_args(keys, args)
-        return [redis_hash.get(str(attribute)) for attribute in attributes]
+        return [redis_hash.get(self._encode(attribute)) for attribute in attributes]
 
     def hset(self, hashkey, attribute, value):
         """Emulate hset."""
 
         redis_hash = self._get_hash(hashkey, 'HSET', create=True)
-        attribute = str(attribute)
+        attribute = self._encode(attribute)
         was_present = attribute in redis_hash
-        redis_hash[attribute] = str(value)
+        redis_hash[attribute] = self._encode(value)
         return 0 if was_present else 1
 
     def hsetnx(self, hashkey, attribute, value):
         """Emulate hsetnx."""
 
         redis_hash = self._get_hash(hashkey, 'HSETNX', create=True)
-        attribute = str(attribute)
+        attribute = self._encode(attribute)
         if attribute in redis_hash:
             return 0
         else:
-            redis_hash[attribute] = str(value)
+            redis_hash[attribute] = self._encode(value)
             return 1
 
     def hincrby(self, hashkey, attribute, increment=1):
@@ -496,9 +500,9 @@ class MockRedis(object):
     def _hincrby(self, hashkey, attribute, command, type_, increment):
         """Shared hincrby and hincrbyfloat routine"""
         redis_hash = self._get_hash(hashkey, command, create=True)
-        attribute = str(attribute)
+        attribute = self._encode(attribute)
         previous_value = type_(redis_hash.get(attribute, '0'))
-        redis_hash[attribute] = str(previous_value + increment)
+        redis_hash[attribute] = self._encode(previous_value + increment)
         return type_(redis_hash[attribute])
 
     def hkeys(self, hashkey):
@@ -526,7 +530,7 @@ class MockRedis(object):
 
         redis_list = self._get_list(key, 'LINDEX')
 
-        if key not in self.redis:
+        if self._encode(key) not in self.redis:
             return None
 
         try:
@@ -570,7 +574,7 @@ class MockRedis(object):
         for key in keys:
             val = pop_func(key)
             if val:
-                return key, val
+                return self._encode(key), val
         return None, None
 
     def blpop(self, keys, timeout=0):
@@ -585,13 +589,13 @@ class MockRedis(object):
         """Emulate lpop."""
         redis_list = self._get_list(key, 'LPOP')
 
-        if key not in self.redis:
+        if self._encode(key) not in self.redis:
             return None
 
         try:
-            value = str(redis_list.pop(0))
+            value = redis_list.pop(0)
             if len(redis_list) == 0:
-                del self.redis[key]
+                self.delete(key)
             return value
         except (IndexError):
             # Redis returns nil if popping from an empty list
@@ -602,21 +606,21 @@ class MockRedis(object):
         redis_list = self._get_list(key, 'LPUSH', create=True)
 
         # Creates the list at this key if it doesn't exist, and appends args to its beginning
-        args_reversed = [str(arg) for arg in args]
+        args_reversed = [self._encode(arg) for arg in args]
         args_reversed.reverse()
-        self.redis[key] = args_reversed + redis_list
+        self.redis[self._encode(key)] = args_reversed + redis_list
 
     def rpop(self, key):
         """Emulate lpop."""
         redis_list = self._get_list(key, 'RPOP')
 
-        if key not in self.redis:
+        if self._encode(key) not in self.redis:
             return None
 
         try:
-            value = str(redis_list.pop())
+            value = redis_list.pop()
             if len(redis_list) == 0:
-                del self.redis[key]
+                self.delete(key)
             return value
         except (IndexError):
             # Redis returns nil if popping from an empty list
@@ -627,14 +631,14 @@ class MockRedis(object):
         redis_list = self._get_list(key, 'RPUSH', create=True)
 
         # Creates the list at this key if it doesn't exist, and appends args to it
-        redis_list.extend(map(str, args))
+        redis_list.extend(map(self._encode, args))
 
     def lrem(self, key, value, count=0):
         """Emulate lrem."""
-        key, value = str(key), str(value)
+        value = self._encode(value)
         redis_list = self._get_list(key, 'LREM')
         removed_count = 0
-        if key in self.redis:
+        if self._encode(key) in self.redis:
             if count == 0:
                 # Remove all ocurrences
                 while redis_list.count(value):
@@ -659,9 +663,9 @@ class MockRedis(object):
                         removed_count += 1
                     else:
                         new_list.append(v)
-                self.redis[key] = list(reversed(new_list))
+                redis_list[:] = list(reversed(new_list))
         if removed_count > 0 and len(redis_list) == 0:
-            del self.redis[key]
+            self.delete(key)
         return removed_count
 
     def ltrim(self, key, start, stop):
@@ -669,7 +673,7 @@ class MockRedis(object):
         redis_list = self._get_list(key, 'LTRIM')
         if redis_list:
             start, stop = self._translate_range(len(redis_list), start, stop)
-            self.redis[key] = redis_list[start:stop + 1]
+            self.redis[self._encode(key)] = redis_list[start:stop + 1]
         return True
 
     def rpoplpush(self, source, destination):
@@ -695,7 +699,7 @@ class MockRedis(object):
         if redis_list is None:
             raise ResponseError("no such key")
         try:
-            redis_list[index] = value
+            redis_list[index] = self._encode(value)
         except IndexError:
             raise ResponseError("index out of range")
 
@@ -720,15 +724,16 @@ class MockRedis(object):
             else:
                 return []
 
+        by = self._encode(by) if by is not None else by
         # always organize the items as tuples of the value from the list itself and the value to sort by
-        if by and '*' in by:
-            items = [(i, self.get(by.replace('*', str(i)))) for i in items]
-        elif by in [None, 'nosort']:
+        if by and b'*' in by:
+            items = [(i, self.get(by.replace(b'*', self._encode(i)))) for i in items]
+        elif by in [None, b'nosort']:
             items = [(i, i) for i in items]
         else:
             raise ValueError('invalid value for "by": %s' % by)
 
-        if by != 'nosort':
+        if by != b'nosort':
             # if sorting, do alpha sort or float (default) and take desc flag into account
             sort_type = alpha and str or float
             items.sort(key=lambda x: sort_type(x[1]), reverse=bool(desc))
@@ -739,11 +744,11 @@ class MockRedis(object):
             if isinstance(get, basestring):
                 # always deal with get specifiers as a list
                 get = [get]
-            for g in get:
-                if g == '#':
+            for g in map(self._encode, get):
+                if g == b'#':
                     results.append([self.get(i) for i in items])
                 else:
-                    results.append([self.get(g.replace('*', str(i[0]))) for i in items])
+                    results.append([self.get(g.replace(b'*', self._encode(i[0]))) for i in items])
         else:
             # if not using GET then returning just the item itself
             results.append([i[0] for i in items])
@@ -771,7 +776,7 @@ class MockRedis(object):
 
         # either store value and return length of results or just return results
         if store:
-            self.redis[store] = results
+            self.redis[self._encode(store)] = results
             return len(results)
         else:
             return results
@@ -796,15 +801,15 @@ class MockRedis(object):
             # we reached the end, back to zero
             result_cursor = 0
         else:
-            result_cursor = str(cursor + count)
+            result_cursor = cursor + count
 
         values = values[cursor:cursor+count]
 
         if match is not None:
-            regex = '^' + match.replace('*', '.*') + '$'
+            regex = re.compile(b'^' + re.escape(self._encode(match)).replace(b'\\*', b'.*') + b'$')
             if not key:
                 key = lambda v: v
-            values = filter(lambda v: re.match(regex, key(v)), values)
+            values = [v for v in values if regex.match(key(v))]
 
         return [result_cursor, values]
 
@@ -849,7 +854,7 @@ class MockRedis(object):
             raise ResponseError("wrong number of arguments for 'sadd' command")
         redis_set = self._get_set(key, 'SADD', create=True)
         before_count = len(redis_set)
-        redis_set.update(map(str, values))
+        redis_set.update(map(self._encode, values))
         after_count = len(redis_set)
         return after_count - before_count
 
@@ -866,7 +871,7 @@ class MockRedis(object):
     def sdiffstore(self, dest, keys, *args):
         """Emulate sdiffstore."""
         result = self.sdiff(keys, *args)
-        self.redis[dest] = result
+        self.redis[self._encode(dest)] = result
         return len(result)
 
     def sinter(self, keys, *args):
@@ -877,7 +882,7 @@ class MockRedis(object):
     def sinterstore(self, dest, keys, *args):
         """Emulate sinterstore."""
         result = self.sinter(keys, *args)
-        self.redis[dest] = result
+        self.redis[self._encode(dest)] = result
         return len(result)
 
     def sismember(self, name, value):
@@ -886,7 +891,7 @@ class MockRedis(object):
         if not redis_set:
             return 0
 
-        result = str(value) in redis_set
+        result = self._encode(value) in redis_set
         return 1 if result else 0
 
     def smembers(self, name):
@@ -897,13 +902,14 @@ class MockRedis(object):
         """Emulate smove."""
         src_set = self._get_set(src, 'SMOVE')
         dst_set = self._get_set(dst, 'SMOVE')
+        value = self._encode(value)
 
         if value not in src_set:
             return False
 
         src_set.discard(value)
         dst_set.add(value)
-        self.redis[src], self.redis[dst] = src_set, dst_set
+        self.redis[self._encode(src)], self.redis[self._encode(dst)] = src_set, dst_set
         return True
 
     def spop(self, name):
@@ -914,7 +920,7 @@ class MockRedis(object):
         member = choice(list(redis_set))
         redis_set.remove(member)
         if len(redis_set) == 0:
-            del self.redis[name]
+            self.delete(name)
         return member
 
     def srandmember(self, name, number=None):
@@ -936,10 +942,10 @@ class MockRedis(object):
             return 0
         before_count = len(redis_set)
         for value in values:
-            redis_set.discard(str(value))
+            redis_set.discard(self._encode(value))
         after_count = len(redis_set)
         if before_count > 0 and len(redis_set) == 0:
-            del self.redis[key]
+            self.delete(key)
         return before_count - after_count
 
     def sunion(self, keys, *args):
@@ -950,7 +956,7 @@ class MockRedis(object):
     def sunionstore(self, dest, keys, *args):
         """Emulate sunionstore."""
         result = self.sunion(keys, *args)
-        self.redis[dest] = result
+        self.redis[self._encode(dest)] = result
         return len(result)
 
     #### SORTED SET COMMANDS ####
@@ -974,7 +980,7 @@ class MockRedis(object):
         # kwargs
         pieces.extend(kwargs.items())
 
-        insert_count = lambda member, score: 1 if zset.insert(str(member), float(score)) else 0
+        insert_count = lambda member, score: 1 if zset.insert(self._encode(member), float(score)) else 0
         return sum((insert_count(member, score) for member, score in pieces))
 
     def zcard(self, name):
@@ -993,7 +999,7 @@ class MockRedis(object):
     def zincrby(self, name, value, amount=1):
         zset = self._get_zset(name, "ZINCRBY", create=True)
 
-        value = str(value)
+        value = self._encode(value)
         score = zset.score(value) or 0.0
         score += float(amount)
         zset[value] = score
@@ -1019,7 +1025,7 @@ class MockRedis(object):
             intersection[member] = reduce(aggregate_func, scores)
 
         # always override existing keys
-        self.redis[dest] = intersection
+        self.redis[self._encode(dest)] = intersection
         return len(intersection)
 
     def zrange(self, name, start, end, desc=False, withscores=False,
@@ -1056,7 +1062,7 @@ class MockRedis(object):
     def zrank(self, name, value):
         zset = self._get_zset(name, "ZRANK")
 
-        return zset.rank(value) if zset else None
+        return zset.rank(self._encode(value)) if zset else None
 
     def zrem(self, name, *values):
         zset = self._get_zset(name, "ZREM")
@@ -1064,10 +1070,10 @@ class MockRedis(object):
         if not zset:
             return 0
 
-        count_removals = lambda value: 1 if zset.remove(value) else 0
+        count_removals = lambda value: 1 if zset.remove(self._encode(value)) else 0
         removal_count = sum((count_removals(value) for value in values))
         if removal_count > 0 and len(zset) == 0:
-            del self.redis[name]
+            self.delete(name)
         return removal_count
 
     def zremrangebyrank(self, name, start, end):
@@ -1080,7 +1086,7 @@ class MockRedis(object):
         count_removals = lambda score, member: 1 if zset.remove(member) else 0
         removal_count = sum((count_removals(score, member) for score, member in zset.range(start, end)))
         if removal_count > 0 and len(zset) == 0:
-            del self.redis[name]
+            self.delete(name)
         return removal_count
 
     def zremrangebyscore(self, name, min_, max_):
@@ -1098,7 +1104,7 @@ class MockRedis(object):
                                                                   start_inclusive=include_start,
                                                                   end_inclusive=include_end)))
         if removal_count > 0 and len(zset) == 0:
-            del self.redis[name]
+            self.delete(name)
         return removal_count
 
     def zrevrange(self, name, start, end, withscores=False,
@@ -1133,12 +1139,16 @@ class MockRedis(object):
         if zset is None:
             return None
 
-        return len(zset) - zset.rank(value) - 1
+        rank = zset.rank(self._encode(value))
+        if rank is None:
+            return None
+
+        return len(zset) - rank - 1
 
     def zscore(self, name, value):
         zset = self._get_zset(name, "ZSCORE")
 
-        return zset.score(value) if zset is not None else None
+        return zset.score(self._encode(value)) if zset is not None else None
 
     def zunionstore(self, dest, keys, aggregate=None):
         union = SortedSet()
@@ -1156,7 +1166,7 @@ class MockRedis(object):
                     union[member] = score
 
         # always override existing keys
-        self.redis[dest] = union
+        self.redis[self._encode(dest)] = union
         return len(union)
 
     #### Script Commands ####
@@ -1245,14 +1255,14 @@ class MockRedis(object):
 
             for i, arg in enumerate(args[3:], 3):
                 # keywords are case-insensitive
-                lower_arg = str(arg).lower()
+                lower_arg = self._encode(arg).lower()
 
                 # handle "limit"
-                if lower_arg == "limit" and i + 2 < len(args):
+                if lower_arg == b"limit" and i + 2 < len(args):
                     start, num = args[i + 1], args[i + 2]
 
                 # handle "withscores"
-                if lower_arg == "withscores":
+                if lower_arg == b"withscores":
                     withscores = True
 
             # do not expect to set score_cast_func
@@ -1279,32 +1289,32 @@ class MockRedis(object):
         """
         Get (and maybe create) a list by name.
         """
-        return self._get_by_type(key, operation, create, 'list', [])
+        return self._get_by_type(key, operation, create, b'list', [])
 
     def _get_set(self, key, operation, create=False):
         """
         Get (and maybe create) a set by name.
         """
-        return self._get_by_type(key, operation, create, 'set', set())
+        return self._get_by_type(key, operation, create, b'set', set())
 
     def _get_hash(self, name, operation, create=False):
         """
         Get (and maybe create) a hash by name.
         """
-        return self._get_by_type(name, operation, create, 'hash', {})
+        return self._get_by_type(name, operation, create, b'hash', {})
 
     def _get_zset(self, name, operation, create=False):
         """
         Get (and maybe create) a sorted set by name.
         """
-        return self._get_by_type(name, operation, create, 'zset', SortedSet(), return_default=False)
+        return self._get_by_type(name, operation, create, b'zset', SortedSet(), return_default=False)
 
     def _get_by_type(self, key, operation, create, type_, default, return_default=True):
         """
         Get (and maybe create) a redis data structure by name and type.
         """
-        key = str(key)
-        if self.type(key) in [type_, 'none']:
+        key = self._encode(key)
+        if self.type(key) in [type_, b'none']:
             if create:
                 return self.redis.setdefault(key, default)
             else:
@@ -1337,7 +1347,7 @@ class MockRedis(object):
         Return a suitable function from (score, member)
         """
         if withscores:
-            return lambda score_member: (score_member[1], score_cast_func(str(score_member[0])))
+            return lambda score_member: (score_member[1], score_cast_func(self._encode(score_member[0])))
         else:
             return lambda score_member: score_member[1]
 
@@ -1384,6 +1394,20 @@ class MockRedis(object):
         if isinstance(score, basestring) and score[0] == '(':
             return False, float(score[1:])
         return True, float(score)
+
+    def _encode(self, value):
+        "Return a bytestring representation of the value. Taken from redis-py connection.py"
+        if isinstance(value, bytes):
+            return value
+        elif isinstance(value, (int, long)):
+            value = str(value).encode('utf-8')
+        elif isinstance(value, float):
+            value = repr(value).encode('utf-8')
+        elif not isinstance(value, basestring):
+            value = str(value).encode('utf-8')
+        else:
+            value = value.encode('utf-8', 'strict')
+        return value
 
 
 def get_total_milliseconds(td):