+++ mockredis/sortedset.py
@@ -34,7 +34,7 @@ class SortedSet(object):
         return len(self._members)
 
     def __contains__(self, member):
-        return str(member) in self._members
+        return member in self._members
 
     def __str__(self):
         return self.__repr__()
@@ -90,7 +90,6 @@ class SortedSet(object):
         """
         Identical to __delitem__, but returns whether a member was removed.
         """
-        member = str(member)
         if member not in self:
             return False
         score = self._members[member]
@@ -104,13 +103,12 @@ class SortedSet(object):
         Identical to __getitem__, but returns None instead of raising
         KeyError if member is not found.
         """
-        return self._members.get(str(member))
+        return self._members.get(member)
 
     def rank(self, member):
         """
         Get the rank (index of a member).
         """
-        member = str(member)
         score = self._members.get(member)
         if score is None:
             return None