+++ CHANGES.md
@@ -1,3 +1,8 @@
+Version 2.9.0.10
+ - Return & store byte strings everywhere (unicode turns into utf-8 by default)
+ - Fix *SCAN returning non-long values.
+ - Fix *TTL returning -1/-2/None - this varies depending on whether StrictRedis is in use.
+
 Version 2.9.0.9
 
  - Support: RENAME and RENAMENX